move: auth 패키지 분리 및 클래스 이동
