+++ src/main/java/soma/ghostrunner/domain/auth/api/AuthController.java
@@ -1,4 +1,4 @@
-package soma.ghostrunner.domain.auth;
+package soma.ghostrunner.domain.auth.api;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
+import soma.ghostrunner.domain.auth.api.dto.SignUpRequest;
 import soma.ghostrunner.domain.auth.application.AuthService;
 
 @Controller
