+++ src/main/java/soma/ghostrunner/domain/auth/application/AuthService.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.server.ResponseStatusException;
-import soma.ghostrunner.domain.auth.AuthIdResolver;
-import soma.ghostrunner.domain.auth.SignUpRequest;
-import soma.ghostrunner.domain.auth.TermsAgreementDto;
+import soma.ghostrunner.domain.auth.resolver.AuthIdResolver;
+import soma.ghostrunner.domain.auth.api.dto.SignUpRequest;
+import soma.ghostrunner.domain.auth.api.dto.TermsAgreementDto;
 import soma.ghostrunner.domain.member.application.MemberService;
 import soma.ghostrunner.domain.member.Member;
 import soma.ghostrunner.domain.member.MemberNotFoundException;
