+++ src/main/java/soma/ghostrunner/domain/auth/resolver/impl/FirebaseUidResolver.java
@@ -1,4 +1,4 @@
-package soma.ghostrunner.domain.auth;
+package soma.ghostrunner.domain.auth.resolver.impl;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthException;
@@ -6,6 +6,7 @@ import com.google.firebase.auth.FirebaseToken;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.stereotype.Component;
+import soma.ghostrunner.domain.auth.resolver.AuthIdResolver;
 
 @Component
 @RequiredArgsConstructor
